+++ src/main/java/com/example/poc/repository/CustomerRepository.java
@@ -1,4 +1,3 @@
-/*
 package com.example.poc.repository;
 
 import com.example.poc.model.entity.Customer;
@@ -6,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
-*/
